+++ uninstall.sh
@@ -6,6 +6,7 @@ set -e
 AGENTDIR="/opt/sfagent"
 TDAGENTCONFDIR="/etc/td-agent-bit"
 ID=`cat /etc/os-release | grep -w "ID" | cut -d"=" -f2 | tr -d '"'`
+SERVICEFILE="/etc/systemd/system/sfagent.service"
 
 uninstall_jcmd()
 {
@@ -24,12 +25,15 @@ uninstall_fluent_bit()
 {
 
 if [ "$ID" = "debian" ] || [ "$ID" = "ubuntu" ]; then
+systemctl stop td-agent-bit
+systemctl disable td-agent-bit
 apt-get purge --auto-remove -y mmdb-bin
 apt-get purge --auto-remove -y td-agent-bit
 fi
 
 if [ "$ID" = "centos" ]; then
 systemctl stop td-agent-bit
+systemctl disable td-agent-bit
 
 yum remove -y td-agent-bit
 rm -rf $TDAGENTCONFDIR
@@ -40,20 +44,25 @@ fi
 
 uninstall_apm_agent()
 {
-
-systemctl stop sfagent
+if [ -f "$SERVICEFILE" ];
+then
+    echo "remove sfagent service"
+    systemctl stop sfagent
+    systemctl disable sfagent
+    rm -f /etc/systemd/system/sfagent.service
+    systemctl daemon-reload
+fi
 rm -rf  $AGENTDIR
 rm -rf /var/log/sfagent
-rm /etc/systemd/system/sfagent.service
 
 }
 
 uninstall_services()
 {
 
+uninstall_apm_agent
 uninstall_fluent_bit
 #uninstall_jcmd
-uninstall_apm_agent
 
 }
